refactor: simplify network classes inheritance
