+++ packages/shared/src/lib/core/network/interfaces/base-network.interface.ts
@@ -1,9 +1,10 @@
 import { IBaseToken } from '@core/token'
 import { Writable } from 'svelte/store'
-import { NetworkHealth, NetworkNamespace } from '../enums'
+import { NetworkHealth, NetworkNamespace, NetworkType } from '../enums'
 import { NetworkId } from '../types'
 
 export interface IBaseNetwork {
+    type: NetworkType
     health: Writable<NetworkHealth>
 
     startStatusPoll(): void
