+++ packages/shared/src/lib/core/network/interfaces/stardust-network.interface.ts
@@ -4,9 +4,7 @@ import { NetworkId } from '../types'
 import { IIscChainConfiguration } from './evm-network-configuration.interface'
 import { IBaseNetwork, IscChain, IStardustNetworkMetadata } from '@core/network'
 
-export interface IStardustNetwork
-    extends Omit<IBaseNetwork, 'type'>,
-        Omit<IStardustNetworkMetadata, 'chainConfigurations'> {
+export interface IStardustNetwork extends IBaseNetwork, Omit<IStardustNetworkMetadata, 'chainConfigurations'> {
     type: NetworkType.Stardust
     namespace: NetworkNamespace.Stardust
     bech32Hrp: string
