+++ packages/shared/src/lib/core/network/types/evm-network.type.ts
@@ -1,0 +1,3 @@
+import { NetworkType } from '../enums'
+
+export type EvmNetworkType = NetworkType.Evm | NetworkType.Isc
