+++ packages/shared/src/lib/core/network/types/index.ts
@@ -1,3 +1,4 @@
+export * from './evm-network.type'
 export * from './network-id.type'
 export * from './network-metadata.type'
 export * from './network.type'
